Use separate class for initial bayeux params
