+++ native/src/main/java/io/ballerinax/salesforce/BasicBayeuxParameters.java
@@ -1,0 +1,65 @@
+/*
+ * Copyright (c) 2026, WSO2 LLC. (http://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.ballerinax.salesforce;
+
+import org.eclipse.jetty.util.ssl.SslContextFactory;
+
+import java.net.URL;
+
+/**
+ * Initial BayeuxParameters used before authentication and endpoint establishment.
+ */
+public class BasicBayeuxParameters implements BayeuxParameters {
+
+    private final String apiVersion;
+    private final SslContextFactory sslContextFactory;
+
+    public BasicBayeuxParameters(String apiVersion) {
+        this.apiVersion = apiVersion;
+        this.sslContextFactory = null;
+    }
+
+    public BasicBayeuxParameters(String apiVersion, SslContextFactory sslContextFactory) {
+        this.apiVersion = apiVersion;
+        this.sslContextFactory = sslContextFactory;
+    }
+
+    @Override
+    public String bearerToken() {
+        throw new IllegalStateException("Have not authenticated");
+    }
+
+    @Override
+    public URL endpoint() {
+        throw new IllegalStateException("Have not established replay endpoint");
+    }
+
+    @Override
+    public String version() {
+        return apiVersion;
+    }
+
+    @Override
+    public SslContextFactory sslContextFactory() {
+        if (sslContextFactory != null) {
+            return sslContextFactory;
+        }
+        return BayeuxParameters.super.sslContextFactory();
+    }
+}
